+++ src/components/layouts/FavList/favlist.styled.js
@@ -1,29 +1,52 @@
 import { styled } from '@mui/material/styles';
-import Box from '@mui/material/Box';
+import { Box, Chip, Paper } from '@mui/material';
 
-const StyledSideNavContainer = styled(Box)`
+export const StyledSideNavContainer = styled(Box)`
   height: 100%;
   width: 350px;
   display: flex;
   flex-direction: column;
 `;
 
-const StyledSideNavHeader = styled(Box)`
+export const StyledSideNavHeader = styled(Box)`
   display: flex;
   justify-content: space-between;
   align-items: center;
   padding: 1rem;
 `;
 
-const StyledSideNavMain = styled(Box)`
+export const StyledSideNavMain = styled(Box)`
   display: flex;
   flex-direction: column;
   padding: 1rem;
   height: 90%;
   justify-content: space-between;
-
   .sidenav-selected-buttons {
   }
 `;
 
-export { StyledSideNavContainer, StyledSideNavHeader, StyledSideNavMain };
+export const StyledBox = styled(Box)`
+  display: flex;
+  flex-direction: column;
+  padding: 1rem;
+  height: 90%
+border:1px solid blue,
+`;
+
+export const StyledPaper = styled(Paper)`
+  maxwidth: 345;
+  position: relative;
+  padding-top: 5px;
+  border: none;
+  div: {
+    height: 100;
+    background: #fff;
+  }
+`;
+export const StyledChip = styled(Chip)`
+  display: block;
+  position: absolute;
+  top: 180px;
+  right: 5px;
+  padding-top: 6px;
+`;